tests/wpt/web-platform-tests/paint-timing/fcp-only/fcp-canvas-context.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP when canvas context is created</title>
</head>
<body>
<script src="../resources/utils.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<canvas id="canvas" width="50" height="50"></canvas>
<script>
  promise_test(async t => {
    assert_precondition(window.PerformancePaintTiming, "Paint Timing isn't supported.");
    await new Promise(r => window.addEventListener('load', r));
    await assertNoFirstContentfulPaint(t);
    const canvas = document.getElementById('canvas');
    const context = canvas.getContext('2d');
    context.fillRect(0, 0, 100, 100);
    await assertFirstContentfulPaint(t);
  }, 'Canvas should count as contentful when context is created');
</script>
</body>

</html>
